Fix handling of rengo teams, remove unnecessary utility file
